07-MCP Client with Multiple Server Support/mcp-client/requirements.txt
langchain
langchain_mcp_adapters 
langgraph 
langchain_google_genai 
google-generativeai 
python-dotenv
mcp